Format fixes. Loading fixes. Need a workaround for index placement(?). It loads, it draws! :]

+++ src/iqm.h
@@ -52,7 +52,7 @@ enum
 };
 struct iqmtriangle
 {
-	glm::detail::tvec3<uint8_t> verts;
+	glm::detail::tvec3<uint> verts;
 };
 struct iqmjoint
 {
